feat: Update Sciantix operations for thermochemistry and fission products
- Updated SetVariablesFunctions.C to include new input variables and history variables for thermochemical data.
- Introduced SetVolatileFPs.C to manage volatile fission products, specifically caesium.
- Added ThermochemistryManifest.C to handle loading and parsing of thermochemistry manifest files.
- Created ThermochemistrySettings.C for loading and parsing thermochemistry settings from a configuration file.
- Enhanced UpdateVariables.C to accommodate new thermochemistry variables and ensure proper updates during simulations.
- Incremented version number to 2.5 and updated authorship information.

+++ include/classes/SciantixArray.h
@@ -8,7 +8,7 @@
 //                                                                                  //
 //  Originally developed by D. Pizzocri & T. Barani                                 //
 //                                                                                  //
-//  Version: 2.2.1                                                                  //
+//  Version: 2.5                                                                    //
 //  Year: 2026                                                                      //
 //  Authors: D. Pizzocri, G. Zullo.                                                 //
 //                                                                                  //
@@ -23,8 +23,6 @@
 #include <map>
 #include <string>
 #include <vector>
-
-#include <type_traits>
 
 /**
  * @class SciantixArray
@@ -75,7 +73,7 @@ template <class T> class SciantixArray
     {
         array = data;
 
-        for (size_t i = 0; i < data.size(); i++)
+        for (int i = 0; i < data.size(); i++)
         {
             map[data[i].getName()] = i;
         }
